Moving TS client to a dedicated subfolder (#274)
