+++ client-ts/Microsoft.AspNetCore.SignalR.Client.TS/Common.ts
