+++ client-ts/Microsoft.AspNetCore.SignalR.Client.TS/Connection.ts
