+++ client-ts/Microsoft.AspNetCore.SignalR.Client.TS/ISignalROptions.ts
