+++ client-ts/Microsoft.AspNetCore.SignalR.Test.Server/EchoEndPoint.cs
