+++ client-ts/Microsoft.AspNetCore.SignalR.Test.Server/Program.cs
