+++ client-ts/Microsoft.AspNetCore.SignalR.Test.Server/Startup.cs
