+++ client-ts/Microsoft.AspNetCore.SignalR.Test.Server/TestHub.cs
