+++ client-ts/Microsoft.AspNetCore.SignalR.Test.Server/web.config
