+++ client-ts/Microsoft.AspNetCore.SignalR.Test.Server/wwwroot/connectionTests.html
