+++ client-ts/Microsoft.AspNetCore.SignalR.Test.Server/wwwroot/js/common.js
