+++ client-ts/Microsoft.AspNetCore.SignalR.Test.Server/wwwroot/js/connectionTests.js
